refactored general settings
