+++ src/pianoroll.py
@@ -1,192 +1,4 @@
-import codecs
-from collections import defaultdict
-from dataclasses import dataclass, field
-from typing import Literal, Optional, Union
-
-from pydantic import BaseModel, Field, computed_field, field_validator, model_validator
-
-INSTRUMENT = str
-BPM = int
-PASS = int
-DEFAULT = -1
-
-# MIDI to Notation
-
-
-@dataclass
-class TimedMessage:
-    time: int
-    type: str
-    note: str = "."
-    cumtime: int = -1
-    duration: int = -1
-
-
-@dataclass
-class TimingData:
-    unit_duration: int
-    units_per_beat: int
-    beats_per_gongan: int
-
-
-# Notation to MIDI
-
-
-@dataclass
-class Note:
-    note: int
-    rest_before: float
-    duration: float
-    rest_after: float
-    description: str
-    symbol: str
-    unicodes: str = ""
-    unicode: int = 0
-    symbol_description: str = ""
-    balifont_symbol_description: str = ""
-
-
-# Metadata
-class Tempo(BaseModel):
-    type: Literal["tempo"]
-    bpm: int
-    passes: list[PASS] = field(default_factory=list)
-    first_beat: Optional[int] = 1
-    beats: Optional[int] = 0
-    passes: Optional[list[int]] = field(default_factory=list)  # On which pass(es) should goto be performed?
-
-    @computed_field
-    @property
-    def first_beat_seq(self) -> int:
-        # Returns the pythonic sequence id (numbered from 0)
-        return self.first_beat - 1
-
-    @model_validator(mode="after")
-    def set_default_pass(self):
-        if not self.passes:
-            self.passes.append(DEFAULT)
-        return self
-
-
-class Label(BaseModel):
-    type: Literal["label"]
-    label: str
-    beat_nr: Optional[int] = 1
-
-    @computed_field
-    @property
-    def beat_seq(self) -> int:
-        # Returns the pythonic sequence id (numbered from 0)
-        return self.beat_nr - 1
-
-
-class GoTo(BaseModel):
-    type: Literal["goto"]
-    label: str
-    beat_nr: Optional[int] | None = None  # Beat number from which to goto. Default is last beat of the system.
-    passes: Optional[list[int]] = field(default_factory=list)  # On which pass(es) should goto be performed?
-
-    @computed_field
-    @property
-    def beat_seq(self) -> int:
-        # Returns the pythonic sequence id (numbered from 0)
-        return self.beat_nr - 1 if self.beat_nr else -1
-
-
-class Gongan(BaseModel):
-    type: Literal["gongan"]
-    kind: str
-
-
-class MetaData(BaseModel):
-    data: Union[Tempo, Label, GoTo, Gongan] = Field(..., discriminator="type")
-
-
-# Flow
-
-
-@dataclass
-class Beat:
-    @dataclass
-    class TempoChange:
-        new_tempo: BPM
-        steps: int = 0
-        incremental: bool = False
-
-    id: int
-    sys_id: int
-    bpm_start: dict[PASS, BPM]  # tempo at beginning of beat (can vary per pass)
-    bpm_end: dict[PASS, BPM]  # tempo at end of beat (can vary per pass)
-    duration: float
-    tempo_changes: dict[PASS, TempoChange] = field(default_factory=dict)
-    staves: dict[INSTRUMENT, list[Note]] = field(default_factory=dict)
-    next: "Beat" = field(default=None, repr=False)
-    goto: dict[PASS, "Beat"] = field(default_factory=dict)
-    _pass_: PASS = 0  # Counts the number of times the beat is passed during generation of MIDI file.
-
-    @computed_field
-    @property
-    def full_id(self) -> str:
-        return f"{int(self.sys_id)}-{self.id}"
-
-    @computed_field
-    @property
-    def sys_seq(self) -> int:
-        # Returns the pythonic sequence id (numbered from 0)
-        return self.sys_id - 1
-
-    def get_bpm_start(self):
-        return self.bpm_start.get(self._pass_, self.bpm_start.get(DEFAULT, None))
-
-    def get_bpm_end(self):
-        return self.bpm_end.get(self._pass_, self.bpm_end.get(DEFAULT, None))
-
-    def get_changed_tempo(self, current_tempo: BPM) -> BPM | None:
-        tempo_change = self.tempo_changes.get(self._pass_, self.tempo_changes.get(DEFAULT, None))
-        if tempo_change and tempo_change.new_tempo != current_tempo:
-            if tempo_change.incremental:
-                return current_tempo + int((tempo_change.new_tempo - current_tempo) / tempo_change.steps)
-            else:
-                return tempo_change.new_tempo
-        return None
-
-
-@dataclass
-class System:
-    # A set of beats.
-    # A System will usually span one gongan.
-    id: int
-    beats: list[Beat] = field(default_factory=list)
-    beat_duration: int = 4
-    gongan: Gongan = field(default_factory=lambda: Gongan(type="gongan", kind="regular"))
-
-
-@dataclass
-class Score:
-    title: str
-    instruments: list[INSTRUMENT]
-    systems: list[System] = field(default_factory=list)
-
-
-@dataclass
-class FlowInfo:
-    labels: dict[str, Beat] = field(default_factory=dict)
-    gotos: dict[str, tuple[System, GoTo]] = field(default_factory=lambda: defaultdict(list))
-
-
-MIDI_TO_COURIER = {
-    36: "\u1ECD",
-    37: "\u1EB9",
-    38: "\u1EE5",
-    39: "\u1EA1",
-    40: "\u0131",
-    41: "o",
-    42: "e",
-    43: "u",
-    44: "a",
-    45: "i",
-}
-VALID_MIDI_MESSAGE_TYPES = ["note_on", "note_off", "rest"]
+from notation_classes import Note
 
 # special note values: -1 = modifies previous note, 0 = rest, 1 = extends previous.
 # key ranges: 0=lower range, 1=mid range, 2=higher range (these denote different frequency ranges per instrument)
@@ -339,57 +151,3 @@ BALIMUSIC4_TO_MIDI = {
     # "ž": Note(note=999, rest_before=0, duration=1, rest_after=0, description='long mutes previous note, quarter duration', symbol='ž', unicode=0x017e, symbol_description='lowercase Z caron', balifont_symbol_description=''),
 }
 # fmt: on
-
-TO_PIANO = {36: 53, 37: 55, 38: 59, 39: 60, 40: 64, 41: 65, 42: 67, 43: 71, 44: 72, 45: 76}
-FROM_PIANO = {53: 36, 55: 37, 59: 38, 60: 39, 64: 40, 65: 41, 67: 42, 71: 43, 72: 44, 76: 45}
-
-
-# ==============================
-NOTENAME = {
-    0: ("dot", "silence (muted note)"),
-    1: ("dash", "open note continuation"),
-    35: ("dotless i underdot", "ding0"),
-    36: ("o underdot", "dong0"),
-    37: ("e underdot", "deng0"),
-    38: ("u underdot", "dung0"),
-    39: ("a underdot", "dang0"),
-    40: ("dotless i", "ding1"),
-    41: ("o", "dong1"),
-    42: ("e", "deng1"),
-    43: ("u", "dung1"),
-    44: ("a", "dang1"),
-    45: ("i (with dot)", "ding2"),
-    46: ("o overdot", "dong2"),
-    47: ("e overdot", "deng2"),
-    48: ("u overdot", "dung2"),
-    49: ("a overdot", "dang2"),
-    60: ("G", "gir"),
-    61: ("W", "gir wadon"),
-    62: ("L", "gir lanang"),
-    63: ("P", "pur"),
-    64: ("T", "tong"),
-    70: ("k", "ke"),
-    71: ("p", "pak"),
-    72: ("d", "dut"),
-    73: ("t", "dut"),
-    74: ("n", "krum"),
-    75: ("u", "pung"),
-    80: ("b", "byong"),
-    81: ("x", "tick"),
-}
-
-
-NOTEDURATION = {
-    (1, 0): ("", " open"),
-    (0.25, 0.75): (" dashed", " long muted"),
-    (0.125, 0.875): (" slashed", " muted"),
-    (0.5, 0): (" single macron", " half duration"),
-    (0.1, 0.4): (" slashed, single macron", " half duration long muted"),
-    (0.25, 0): (" double macron", " quarter duration"),
-    (0.05, 0.2): (" slashed, double macron", " quarter duration muted "),
-}
-
-if __name__ == "__main__":
-    for key, note in BALIMUSIC4_TO_MIDI.items():
-        # print(f"{note.symbol} - {note.description} - {note.balifont_symbol_description}")
-        print(f"{note.symbol}  - {note.description} - {codecs.encode(note.symbol, encoding='unicode-escape')}")
